oracle compliance

+++ codebase/projects/core/db/bering/021_extension_iteration8/009_drop_column_disease_history_anatomic_site_id.groovy
@@ -1,8 +1,15 @@
 class DropColumnDiseaseHistoryAnatomicSiteId extends edu.northwestern.bioinformatics.bering.Migration {
     void up() {
-    	 execute("alter table disease_history drop column anatomic_site_id");
-    	 execute("drop table anatomic_sites cascade");
-    	 execute("drop sequence anatomic_sites_id_seq cascade");
+    	execute("alter table disease_history drop column anatomic_site_id");
+    	 
+    	if (databaseMatches('oracle')) {
+		   	execute("drop table anatomic_sites");
+		   	execute("drop sequence ANATOMIC_SITES_ID_SEQ");
+	 	}
+	 	if (databaseMatches('postgres')) {
+		   	execute("drop table anatomic_sites cascade");
+		   	execute("drop sequence anatomic_sites_id_seq cascade");
+	 	}
     } 
 
 	void down() {
